requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-humanize
djangorestframework
EasyProcess
entrypoint2
filelock
humanize
Markdown
MouseInfo
mss
pillow
pipenv
platformdirs
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
setuptools
sqlparse
tzdata
virtualenv